+++ lib/flutter_view_controller.dart
@@ -3,4 +3,5 @@ library flutter_view_controller;
 export 'src/notifier.dart';
 export 'src/plug.dart';
 export 'src/view.dart';
+export 'src/screen_size.dart';
 export 'src/navigator_monitor.dart';